Preliminary changes

+++ ToyWorld/ToyWorldAPI/Render/IAvatarRenderRequests.cs
@@ -11,7 +11,7 @@ namespace GoodAI.ToyWorld.Control
         /// <summary>
         /// 
         /// </summary>
-        float AvatarID { get; }
+        int AvatarID { get; }
 
 
         /// <summary>
